Move FromStr impl for NetAddr from crate root to netaddr mod
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>
