+++ src/netaddr.rs
@@ -1,5 +1,7 @@
+use crate::netaddr_error::NetAddrError;
 use crate::netv4addr::Netv4Addr;
 use crate::netv6addr::Netv6Addr;
+use core::str::FromStr;
 use std::net::IpAddr;
 
 /// A structure representing an IP network.
@@ -24,6 +26,21 @@ impl NetAddr {
 		match self {
 			NetAddr::V4(v4) => IpAddr::V4(*v4.addr()),
 			NetAddr::V6(v6) => IpAddr::V6(*v6.addr()),
+		}
+	}
+}
+
+impl FromStr for NetAddr {
+	type Err = NetAddrError;
+
+	fn from_str(string: &str) -> Result<Self, NetAddrError> {
+		let as_v4: Result<Netv4Addr, NetAddrError> = string.parse::<Netv4Addr>();
+		let as_v6: Result<Netv6Addr, NetAddrError> = string.parse::<Netv6Addr>();
+
+		match (as_v4, as_v6) {
+			(Ok(v4), _) => Ok(NetAddr::V4(v4)),
+			(_, Ok(v6)) => Ok(NetAddr::V6(v6)),
+			(Err(_e4), Err(e6)) => Err(e6),
 		}
 	}
 }
